refactor: move todo-mock.ts to api/mocks
